+++ packages/fxa-profile-server/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.144.4
+
+No changes.
+
+## 1.144.3
+
+No changes.
+
 ## 1.144.2
 
 No changes.