+++ packages/fxa-shared/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change history
 
+## 1.144.4
+
+No changes.
+
+## 1.144.3
+
+No changes.
+
 ## 1.144.2
 
 No changes.